LayoutTests/http/tests/quicklook/hide-referer-on-navigation.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/tap-at-point-and-notify-done.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<p>This test verifies that the HTTP referrer is hidden when navigating from a Microsoft Word QuickLook preview. This test PASSED only if nothing is printed below.</p>
<iframe src="resources/hide-referer-on-navigation.docx" onload="tapAtPoint(this.offsetLeft + 5, this.offsetTop + 5, function () { /* Do nothing */ })"></iframe>
</body>
</html>
